MySales/salesApp/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Deprecated==1.2.14
Django==4.2.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.15.0
